+++ BE/src/main/java/dev/codesquad/airbnb02/application/dto/RoomDetailResponseDto.java
@@ -3,7 +3,7 @@ package dev.codesquad.airbnb02.application.dto;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import dev.codesquad.airbnb02.domain.room.entity.Image;
+import dev.codesquad.airbnb02.domain.room.vo.Image;
 import dev.codesquad.airbnb02.domain.room.entity.Room;
 import dev.codesquad.airbnb02.domain.room.vo.Locale;
 import lombok.Builder;
